Fix H5FDdsmInt32/64 conversion since both MPI and socket can only send int size messages - Add WIN32 mutex - Fix H5TOOLS_DLL export

+++ Testing/H5FDdsmReceiverTest1.cxx
@@ -32,11 +32,11 @@ int main (int argc, char* argv[])
 {
   int provided, rank, size;
   MPI_Comm dcomm = MPI_COMM_WORLD;
-  double DSMSize = 16;
+  long DSMSize = 16;
   int commType = H5FD_DSM_COMM_SOCKET;
 
   if (argv[1]) {
-    DSMSize = atof(argv[1]);
+    DSMSize = atol(argv[1]);
   }
 
   if (argv[2]) {
